requirements.txt
six
Faker
pillow
markdown
python-dateutil

django
django-environ
django-filter
django-allauth == 0.55.0
django-resized
python-decouple
django-crispy-forms
folium
django-tinymce

django-browser-reload
